Cancel underlying response handler when grpc call is cancelled
Prior to this commit, we would race `finalHandler` with the completed
task of the `isCancelled` promise but we would not cancel the loser.
Therefore, the response handler would not cancel the underlying task and
grpc resources would not be freed up.

By replacing `racePair` with `race`, we automatically cancel the loser
fiber and free up resources as `responseHandler` will propagate the
cancellation up the task/obsevable chain as expected.