Replace `OverflowStrategy.Fail(4)` with unbounded
There is no use to specifying an arbitrary number such as 4 to stop any
kind of excessive buffering in our subject. If backpressure works well,
then that will never be necessary, especially taking into account that
we have a publish subject that doesn't buffer. If backpressure doesn't
work well, then setting a limit to 4 will not buy us anything.

It's better to use an unbounded overflow strategy and just invest in our
test infrastructure to make sure that backpressure works.

We will most likely revisit this change when prefetching is implemented
in favor of another overflow strategy such `BackPressure(n)` though
still more research is required.