Guard `sendMessage` with `isReady` following grpc-java advice
As specified in the grpc-java documentation [1], the `onReadyHandler`
can be called spuriously by the grpc framework while the grpc call is
still not ready and its buffer hasn't been drained.

Because of that, we can't just wait for the async var to be full to send
a message. We need to check that the call is indeed `true`, otherwise we
could overflow the netty buffer. If the call isn't `true` yet, we can
safely assume that the on ready handler will be called again, the async
var will again be full and the `isReady` check will eventually yield
a positive value. After that, the message will be sent.

[1]: https://github.com/grpc/grpc-java/pull/5799/files