Add initial back pressure support (#2)
* Pr review comments

* adding a backpressure example with ready

* Update waitUntilReady call

Update waitUntilReady call

Co-authored-by: Jorge <jorge@vican.me>

* Implement backpressure with on Ready on all calls.

* Add empty test cases

* Fix tests

* Chunking messages.
Testing backpressure (messy)
Backpressure is stuck.

* Add backpressure to each call.

* Disable timout test

* Add tests for ClientCallListeners

* Add tests for Backpressure

* Fixed back pressure

* Corrected the subject

* Corrected the subject

* Downgrade Scala 2.12 and 2.13 versions to work with Metals

Metals doesn't yet support Scala 2.12.13 and 2.13.5.

* Improve client call listeners semantics and abstract away logic

- Abstract away requesting messages from unary call for + readability
- Isolate response observables correctly to propagate the context at
  subscription time without any task local interferences
- Abstract away sending requests to the server and make sure that if
  there is an error while receiving messages we cancel the sending fiber
- Revert some unimportant stylistic changes

* Scope new dependencies to test project in build.sbt

- These dependencies were being added to every downstream user when they
  should only be scoped to the test project
- The test suite defined in `ClientCallListenersTest` wasn't running
  because of the same issue

* Simplify server call handlers

* Improve tests for ClientCallListeners

* Add tests for StreamingCallListener

* Improved TestBackpressure by increasing the delay.
Backpressure doesn't work on server side because it is always send directly.

* Use inprocessServer to avoid the netty buffering.

* Fix failing ServerCallHandlersTest test.

* Made server handler use a plubish model instead of a limited replay.
Fixed tests by adding usePlaintext() on the client.

* Made all calls use a publish subject.

* Made calls using the replay limited again because otherwise the calls become unstable.

* Fix tests

* Remove copyright

* Fix the order of start call and the interaction with the call
Add tests to verify the order in the ClientCall

* Fix deprecation warning

* Fix flaky test

* Don't format comments

The previous version is the default pattern used throughout the codebase
so let's keep it consistent with what there is.

* Add small formatting improvements

* Replace `OverflowStrategy.Fail(4)` with unbounded

There is no use to specifying an arbitrary number such as 4 to stop any
kind of excessive buffering in our subject. If backpressure works well,
then that will never be necessary, especially taking into account that
we have a publish subject that doesn't buffer. If backpressure doesn't
work well, then setting a limit to 4 will not buy us anything.

It's better to use an unbounded overflow strategy and just invest in our
test infrastructure to make sure that backpressure works.

We will most likely revisit this change when prefetching is implemented
in favor of another overflow strategy such `BackPressure(n)` though
still more research is required.

* Move flow control to call sites instead of `incomingResponses`

No need to abstract away these bits. Repetition in the call sites helps
readability and understand how we're dealing with flow control right at
the spot where we also start the call and send messages.

* Apply stylistic changes to client and server calls

* Move `ClientCallMock` to independent Scala file

Just like `ServerCallMock`.

* Remove Copyright headers from the source

* Move `ServerCallMock` to test utils package

* Change name of server unit tests and simplify test descriptions

* Simplify test names

* Guard `sendMessage` with `isReady` following grpc-java advice

As specified in the grpc-java documentation [1], the `onReadyHandler`
can be called spuriously by the grpc framework while the grpc call is
still not ready and its buffer hasn't been drained.

Because of that, we can't just wait for the async var to be full to send
a message. We need to check that the call is indeed `true`, otherwise we
could overflow the netty buffer. If the call isn't `true` yet, we can
safely assume that the on ready handler will be called again, the async
var will again be full and the `isReady` check will eventually yield
a positive value. After that, the message will be sent.

[1]: https://github.com/grpc/grpc-java/pull/5799/files

Co-authored-by: Jorge <jorge@vican.me>
Co-authored-by: Jorge Vicente Cantero <jorgevc@fastmail.es>