Improve error messages for cancellations
We are getting some unexpected cancellation rpc errors that have a
`null` cause, that is the original error on the client side is not
getting propagated from the client to the server. These changes hope to
make every call cancellation clearer to help identify where the
cancelled status is coming from.