Add client cause to cancelled server call with null cause
When a call that sends many requests is established, we listen first for the error
on the server side than the client side even if the one that originated
the cancellation is the server. This can cause the problem that
downstream users see an empty cancelled error with no context related to
the client error and it can cause confusion. This was also making some
of our tests fail spuriously because of a race when completing the
cancellation exception context.

To fix this, we add some logic to the client-side of a bidirectional and
a client streaming call to propagate any exception that has been thrown
on the fiber sending the client requests if no cause is already present.