Make several changes in e2e test setup (#12)
* Change test utilities not to use fixtures

In principle, it's best we don't want to share state across different
test cases. For that reason, fixtures are not appropriate for testing
each case because they run at the beginning and end of test suites.

* Consolidate tests for `TestService` and backpressure specs

- Optimize test logic for simplicity and readability
- Prefer exhaustiveness when checking status runtime exceptions
- Use diffs to test expected behaviors instead of value comparisons
- Don't use `ReplaySubject` because replay subject caching doesn't
  simulate real-world scenarios accurately, use `PublishSubject` instead
  and add small random variances when sending messages

* Remove unused `GrpcServerFixture` logic

* Improve error messages for cancellations

We are getting some unexpected cancellation rpc errors that have a
`null` cause, that is the original error on the client side is not
getting propagated from the client to the server. These changes hope to
make every call cancellation clearer to help identify where the
cancelled status is coming from.

* Add client cause to cancelled server call with null cause

When a call that sends many requests is established, we listen first for the error
on the server side than the client side even if the one that originated
the cancellation is the server. This can cause the problem that
downstream users see an empty cancelled error with no context related to
the client error and it can cause confusion. This was also making some
of our tests fail spuriously because of a race when completing the
cancellation exception context.

To fix this, we add some logic to the client-side of a bidirectional and
a client streaming call to propagate any exception that has been thrown
on the fiber sending the client requests if no cause is already present.

* Fix wrong time unit from seconds to milliseconds