Add server call options API + support server-side buffering
A similar adjustment to the one made for the client call options API.
I've taken the libery to redesign how server call options are passed in
regardless of binary compatibility as we don't promise backwards
compatibility until we make a stable release >= 1.0.

I have added some more documentation to the server call to make it clear
that we are trying to replicate the API in the underlying grpc java
library.